Base is complete Needs Sorting ability
Everything around the base ideas of playing games and such are
complete. A Sorting ability still needs to be implemented such that it
can be read easier, I will also need to implement a save function so
that it can save to a file. Finally, different sets of matches such as
tournaments need to be implemented